サブイベントでしか使っていない定数値を近くに移動
